# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://ecommerce.local:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "weight of product",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "250"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "store a new product",
                "description": "store a new product",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{product}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "update product details",
                "description": "update product request",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product id of product",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "StoreProductRequest": {
                "title": "Store Product Reqeust",
                "description": "Store Product request body Data",
                "required": [
                    "name",
                    "weight"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of the product",
                        "example": "test"
                    },
                    "rating": {
                        "title": "rating",
                        "description": "rating of the product",
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the product",
                        "type": "string",
                        "example": "test"
                    },
                    "weight": {
                        "title": "weight",
                        "description": "weight of the product",
                        "example": "100"
                    },
                    "brand_id": {
                        "title": "brand_id",
                        "description": "brand_id of the product",
                        "example": "1"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id of the product",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "title": "Update Product Reqeust",
                "description": "Update Product request body Data",
                "required": [
                    "name",
                    "weight"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of the product",
                        "example": "test"
                    },
                    "rating": {
                        "title": "rating",
                        "description": "rating of the product",
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the product",
                        "type": "string",
                        "example": "test"
                    },
                    "weight": {
                        "title": "weight",
                        "description": "weight of the product",
                        "example": "100"
                    },
                    "brand_id": {
                        "title": "brand_id",
                        "description": "brand_id of the product",
                        "example": "1"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id of the product",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}